JAMES-2927 Prevent unwanted tumbstone creation for ExecutionDetails projection
A tumbstone was created when a null value is specified in a prepared
statement.

This basically happened several time a "task details" get saved - as a
task cannot be both failed and completed.

This is due to the fact that null has the meaning `remove` and not the meaning `unspecified`, which is represented by no binding at all.

Of course unwanted tumbstones occurs with a performance cost.

The recommended method for fixing on the latest version of cassandra is to not bind the null value.

Read this for further information: https://thelastpickle.com/blog/2016/09/15/Null-bindings-on-prepared-statements-and-undesired-tombstone-creation.html